Reach and avoid specs allowed
Reach and avoid specifications are abled to be passed in through files now. The obs_g is now an array of goals, with the added parameter 'reach' that is true if the object is a goal, and false when it should be avoided.

Currently, the reach/avoid spec is optimized poorly as a +/-, soon to be fixed.